+++ src/schema.js
@@ -149,6 +149,12 @@ const attributes = {
     basic_auth: {
         parents: ['backend']
     },
+    beta_scope: {
+        parents: ['api', 'endpoint']
+    },
+    beta_scope_claim: {
+        parents: ['jwt']
+    },
     body: {
         parents: ['request', 'response']
     },